requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.1.0
Werkzeug==2.3.7
python-dotenv==1.0.1
email-validator==2.1.1
emoji==2.14.1
WTForms==3.0.1
gunicorn==21.2.0 